style: createEditMenuAction
